Implement personal appointment save logic

+++ AppointmentsService/AppointmentsService/src/main/java/com/fotoalpha/appointmentsservice/ResponseRequest/SaveOwnMadeWeddingAppointmentRequest.java
@@ -19,5 +19,5 @@ public class SaveOwnMadeWeddingAppointmentRequest {
     LocalDate orderDate;
     LocalDate appointmentDate;
     LocalTime appointmentTime;
-    List<Integer> eventIds;
+    List<Long> eventIds;
 }
